+++ app/Http/Controllers/Mentors/MentorController.php
@@ -29,10 +29,18 @@ class MentorController extends Controller
                 "message"=>"Unauthorized"
             ],401);
         }
+        $mentor=Mentor::find($id);
+        if(!$mentor){
+            return response()->json([
+                "status"=>404,
+                "message"=>"Not Found"
+            ],404);
+        }
         $mentor=Mentor::leftjoin("groups","groups.id","=","mentors.group_id")
         ->leftjoin("interns","interns.mentor_id","=","mentors.id")
         ->where("mentors.id","$id")
         ->get(["mentors.name","mentors.surname","mentors.city","mentors.skype","interns.name as intern_name","interns.surname as intern_surname","groups.title as group_title"]);
+
         return response()->json([
             "status"=>200,
             "data"=>$mentor
@@ -46,9 +54,9 @@ class MentorController extends Controller
             ],401);
         }
         $attributes = $request->validate([
-            "name"=>["string","max:255","alpha"],
-            "surname"=>["string","max:255"],
-            "city"=>["string","max:255"],
+            "name"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
+            "surname"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
+            "city"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
             "skype"=>["string","max:255"],
             "email"=>["required","max:255","email"],
             "password"=>["required","min:6","string"],
@@ -75,29 +83,29 @@ class MentorController extends Controller
             "data"=>$mentor
         ],201);
     }
-    public function update(Request $request, Mentor $mentor ){
+    public function update(Request $request, $id){
         if(!$request->header('Authorization')){
             return response()->json([
                 "status"=>401,
                 "message"=>"Unauthorized"
             ],401);
         }
+        $mentor=Mentor::find($id);
+        if(!$mentor){
+            return response()->json([
+                "status"=>404,
+                "message"=>"Not Found"
+            ],404);
+        }
         $attributes = $request->validate([
-            "name"=>["string","max:255"],
-            "surname"=>["string","max:255"],
-            "city"=>["string","max:255"],
+            "name"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
+            "surname"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
+            "city"=>["string","max:255","regex:/^[a-zA-Z\s]*$/"],
             "skype"=>["string","max:255"],
             "email"=>["max:255","email"],
             "password"=>["min:6","string"],
             "group_id"=>["numeric"],
         ]);
-        $user=Mentor::where("email",$attributes["email"])->first();
-        if($user){
-            return response()->json([
-                "status"=>403,
-                "message"=>"Already exists"
-            ],403);
-        }
         if(!$attributes){
             return response()->json([
                 "status"=>422,
@@ -119,12 +127,19 @@ class MentorController extends Controller
             "data"=>$mentor
         ],200);
     }
-    public function destroy( Mentor $mentor, Request $request){
+    public function destroy( Request $request,$id){
         if(!$request->header('Authorization')){
             return response()->json([
                 "status"=>401,
                 "message"=>"Unauthorized"
             ],401);
+        }
+        $mentor=Mentor::find($id);
+        if(!$mentor){
+            return response()->json([
+                "status"=>404,
+                "message"=>"Not Found"
+            ],404);
         }
         $mentor->delete();
         return response()->json([
